// src/interfaces/organisations.ts
export interface YPEOrg {
    country: string;
    name: string;
    website?: string;
    facebook?: string;
    instagram?: string;
    twitter?: string;
    mastodon?: string;
    tiktok?: string;
    youtube?: string;
    image: string,
    logoAspectRation: string
    email?: string
    active: boolean
}